backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.8.0
httpx>=0.27.0
python-multipart>=0.0.9
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
bibtexparser>=1.4.1
numpy>=1.26.0
notion-client>=2.2.1
python-dateutil>=2.9.0
python-dotenv>=1.0.0
